write meta-merging program
merge-meta.rkt is a racket program to merge metadata from a list of
directories. In actuality, it keeps the most recent versions of the
relevant information; crucially, information that is deleted in newer
metadatas is preserved.

This needs to finish being integrated into the build system, however,
which also needs to take into account the new metadata format.