config: simplify define-dynamic-definer macro
First, some definitions:
- a (syntax) binding to a syntax-transformer is a macro
- a compile-time function that produces a syntax-transformer can be used
  to implement a macro
- a syntax-parser that returns a syntax object is a syntax-transformer

Thus, a compile-time function that uses a syntax-parser that returns a
syntax object can be used to implement a macro, when a binding is made
to the result.

This is particularly useful when defining macros that define macros:
without compile-time helpers, the inner define-syntax (and myriad
others, e.g., define-syntax-parse-rule) become confusing. Not only is it
difficult to keep track of which bits come from where, but any
non-trivial templating may require an extra (...) form, as in the
previous commit, which used to (... (~? default)) to make sure the
conditional splice happened during the inner macro (and NOT the outer
macro).

With compile-time helpers, however, the inner macros are implemented by
a call to the helper (with any relevant state) and bound as a
syntax-transformer. The helpers can make use of (e.g.) syntax-parser to
drive the implementation and syntax-quasiquote/syntax-unquote (#`/#,) to
mix-in parameterized state. In many cases this version is also more
efficient.

Now define-dynamic-definer should make sense: it expands to a function
definition and a syntax-transformer definition. The syntax-transformer
is computed by a compile-time helper function that uses syntax-parser to
implement a syntactic transformation. It splices in the parameter name
(bound to a syntax-object in the template where it is called) to call
the correct function.

If you are surprised that there is no (require (for-syntax
syntax/parse)), note that syntax/parse/define provides for-syntax all of
syntax/parse.