include full date with formatted time
Since this rewrite also includes the "split-by-date" feature (each
channel has its messages spread across multiple pages, to avoid one long
page not loading or consuming too much browser memory), we include the
full date & time in the message. It is also available in the web
address, but that may not be the most accessible for everyone.