get-channel-purpose: lookup the converted symbol
A previous renaming switched from channel-id to channel-name to re-align
with the Jekyll/ruby code. I apparently forgot to update the let
binding.

TODO: tests so this stuff doesn't happen.

Aside: the let-over-define looks weird, doesn't it? I'm pretty sure this
is effectively required (unless you want to use `let*`) because of this:

```racket
(define (f x)
  (define x (add1 x))
  ;; more defines…
  (* x 2))
(f 1)
; x: undefined;
;  cannot use before initialization
```

`define` permits recursion, which is useful, but it means the
right-hand-side `x` (`channel-name`) cannot be used to refer to the
outer binding, only to the `define`d binding or other
inside-the-`define` bindings. This is not true of `let`, where the
binding not even available until after the expression has been evaluated
(partly as a semantic decision, partly as a consequence of the typical
`let` expansion to `lambda`).