correct txexpr[*] usage
- `txexpr*` takes a series of elements (a rest parameter)
- `txexpr` takes a list of elements
- `map` produces a list

Ergo, `txexpr*` is the wrong choice here, unless we want to also include
`apply`, which seems silly.

While we're here: why `map` everywhere instead of `for/list`? Well, I
can't be sure, but it looks like the `for/list` expansion doesn't
compile down to something resembling a `map` when it can in simple cases
like these. And I'm assuming that `map` is straightforward. So for me,
it's readable and performant. For more complex cases, `for/list` might
be the right tool; but with only lists and a function over their
elements, I'll stick to map. (cf. use of `for/async`.)