render: use system instead of raco
After hours debugging, I discovered that the parallel render uses
places, which do not inherit the parameters current-directory,
current-project-root, or really any parameters. Thus the constant
crashes and serialization (which lacks this problem).

In order to continue benefiting from parallelism, we'll use system
wrapped in a change to the current-directory.