template: forego when/splice
At least for beeswax, things don't quite work out correctly. We get the
literal '(@) in documents, or a series of odd element strings that
break the HTML. If we used pollen notation exclusively in the body, we
could probably call ->html on it and be done. But I'm lazy, so just use
some empty strings.