do more setup in parallel
The current implementation uses futures (for/async); I'm considering
using places where it might make a difference.

I'm also considering trying a big refactor: the new design would
(hypothetically)
- do the unpacking as before,
- copy required configuration files (index.ptree, pollen.rkt,
  template.html.rkt, &c.) to pollen,
- in parallel places, convert & render _data files to pollen files to
  output files as they are ready.

  This looks like having a queue of the files. Each place grabs
  one (or a batch of 10, or whatever) off the queue, converts them, and
  renders them. Repeat until queue is empty.

I'm not sure if this would actually make things faster or not, but it is
closer in concept to "streaming" than "batch" computation, perhaps.